+++ lib/dry/cli.rb
@@ -83,12 +83,13 @@ module Dry
     # Invoke the CLI
     #
     # @param arguments [Array<string>] the command line arguments (defaults to `ARGV`)
-    # @param out [IO] the standard output (defaults to `$stdout`)
-    # @param err [IO] the error output (defaults to `$stderr`)
+    # @param stderr [IO] the error output (defaults to `$stderr`)
+    # @param stdin [IO] the standard input (defaults to `$stdin`)
+    # @param stdout [IO] the standard output (defaults to `$stdout`)
     #
     # @since 0.1.0
-    def call(arguments: ARGV, out: $stdout, err: $stderr)
-      @out, @err = out, err
+    def call(arguments: ARGV, stderr: $stderr, stdin: $stdin, stdout: $stdout)
+      @stderr, @stdin, @stdout = stderr, stdin, stdout
       kommand ? perform_command(arguments) : perform_registry(arguments)
     rescue SignalException => exception
       signal_exception(exception)
@@ -108,11 +109,15 @@ module Dry
 
     # @since 0.6.0
     # @api private
-    attr_reader :out
+    attr_reader :stderr
+
+    # @since unreleased
+    # @api private
+    attr_reader :stdin
 
     # @since 0.6.0
     # @api private
-    attr_reader :err
+    attr_reader :stdout
 
     # Invoke the CLI if singular command passed
     #
@@ -123,10 +128,6 @@ module Dry
     # @api private
     def perform_command(arguments)
       command, args = parse(kommand, arguments, [])
-
-      command.instance_variable_set(:@err, err) unless command.instance_variable_defined?(:@err)
-      command.instance_variable_set(:@out, out) unless command.instance_variable_defined?(:@out)
-
       command.call(**args)
     end
 
@@ -142,10 +143,7 @@ module Dry
       return spell_checker(result, arguments) unless result.found?
 
       command, args = parse(result.command, result.arguments, result.names)
-      return err.puts(Usage.call(result)) unless command.respond_to?(:call)
-
-      command.instance_variable_set(:@err, err) unless command.instance_variable_defined?(:@err)
-      command.instance_variable_set(:@out, out) unless command.instance_variable_defined?(:@out)
+      return stderr.puts(Usage.call(result)) unless command.respond_to?(:call)
 
       result.before_callbacks.run(command, **args)
       command.call(**args)
@@ -179,28 +177,31 @@ module Dry
     # @since 0.6.0
     # @api private
     def build_command(command)
-      command.is_a?(Class) ? command.new : command
+      return command unless command.is_a?(Class)
+      return command.new(stderr: stderr, stdin: stdin, stdout: stdout) if CLI.command?(command)
+
+      command.new
     end
 
     # @since 0.6.0
     # @api private
     def help(command, prog_name)
-      out.puts Banner.call(command, prog_name)
+      stdout.puts Banner.call(command, prog_name)
       exit(0) # Successful exit
     end
 
     # @since 0.6.0
     # @api private
     def error(result)
-      err.puts(result.error)
+      stderr.puts(result.error)
       exit(1)
     end
 
     # @since 1.1.1
     def spell_checker(result, arguments)
       spell_checker = SpellChecker.call(result, arguments)
-      err.puts "#{spell_checker}\n\n" if spell_checker
-      err.puts Usage.call(result)
+      stderr.puts "#{spell_checker}\n\n" if spell_checker
+      stderr.puts Usage.call(result)
       exit(1)
     end